+++ server.js
@@ -50,8 +50,8 @@ function withAppKeyCheck (cb) {
 
 app.use('/auth/me', withAppKeyCheck(sendJsonData(200, currenUser)));
 
-app.use('/auth/login/twitter', sendRedirect('/learningpath/login/failure'));
-app.use('/auth/login', sendRedirect('/learningpath/login/success/' + AUTH_TOKEN));
+app.use('/auth/login/twitter', sendRedirect('/login/failure'));
+app.use('/auth/login', sendRedirect('/login/success/' + AUTH_TOKEN));
 
 app.use('/paths/private', withAppKeyCheck(sendJsonData(200, data.private)));
 app.use('/paths', sendJsonData(200, data.public));